+++ apps/brukerdialog-sanity/schemaTypes/meldInnEndring/meldInnEndringMeldFraSchemaType.ts
@@ -1,0 +1,3 @@
+import { defineType } from 'sanity'
+
+export const meldInnEndringMeldFraSchemaType = defineType({})
